refactor: userフィーチャーをトップレベルパッケージに移動
domain.user, application.UserService, infrastructure.repository.JdbcUserRepository,
presentation.user をすべて user/ 配下に再配置し、
関連する全ファイルの import を更新。
