+++ backend/src/integrationTest/java/com/youmorry/expensetracker/integration/TransactionSearchIntegrationTest.java
@@ -5,7 +5,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.youmorry.expensetracker.domain.user.UserId;
+import com.youmorry.expensetracker.user.domain.UserId;
 import org.junit.jupiter.api.Test;
 
 /** 支出検索フィルタの一気通貫テスト。 */
