+++ backend/src/integrationTest/java/com/youmorry/expensetracker/integration/UserIntegrationTest.java
@@ -9,7 +9,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.youmorry.expensetracker.domain.user.UserId;
+import com.youmorry.expensetracker.user.domain.UserId;
 import org.junit.jupiter.api.Test;
 
 /** ユーザーエンドポイントの統合テスト。 */
