+++ backend/src/main/java/com/youmorry/expensetracker/application/auth/AuthService.java
@@ -1,7 +1,7 @@
 package com.youmorry.expensetracker.application.auth;
 
-import com.youmorry.expensetracker.domain.user.User;
-import com.youmorry.expensetracker.domain.user.UserRepository;
+import com.youmorry.expensetracker.user.domain.User;
+import com.youmorry.expensetracker.user.domain.UserRepository;
 import java.util.Objects;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
