+++ backend/src/main/java/com/youmorry/expensetracker/infrastructure/repository/JdbcTransactionRepository.java
@@ -3,7 +3,7 @@ package com.youmorry.expensetracker.infrastructure.repository;
 import com.youmorry.expensetracker.domain.transaction.Transaction;
 import com.youmorry.expensetracker.domain.transaction.TransactionId;
 import com.youmorry.expensetracker.domain.transaction.TransactionRepository;
-import com.youmorry.expensetracker.domain.user.UserId;
+import com.youmorry.expensetracker.user.domain.UserId;
 import java.util.List;
 import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
