+++ backend/src/main/java/com/youmorry/expensetracker/shared/infrastructure/persistence/IdConverters.java
@@ -2,7 +2,7 @@ package com.youmorry.expensetracker.shared.infrastructure.persistence;
 
 import com.youmorry.expensetracker.domain.category.CategoryId;
 import com.youmorry.expensetracker.domain.transaction.TransactionId;
-import com.youmorry.expensetracker.domain.user.UserId;
+import com.youmorry.expensetracker.user.domain.UserId;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.data.convert.ReadingConverter;
 import org.springframework.data.convert.WritingConverter;
