+++ backend/src/main/java/com/youmorry/expensetracker/user/application/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.youmorry.expensetracker.user.application;
+
+import org.jspecify.annotations.NullMarked;
