+++ backend/src/main/java/com/youmorry/expensetracker/user/domain/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.youmorry.expensetracker.user.domain;
+
+import org.jspecify.annotations.NullMarked;
