+++ backend/src/main/java/com/youmorry/expensetracker/user/infrastructure/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.youmorry.expensetracker.user.infrastructure;
+
+import org.jspecify.annotations.NullMarked;
