+++ backend/src/main/java/com/youmorry/expensetracker/user/package-info.java
@@ -1,4 +1,4 @@
 @NullMarked
-package com.youmorry.expensetracker.domain.user;
+package com.youmorry.expensetracker.user;
 
 import org.jspecify.annotations.NullMarked;
