+++ backend/src/main/java/com/youmorry/expensetracker/user/presentation/UserController.java
@@ -1,8 +1,8 @@
-package com.youmorry.expensetracker.presentation.user;
+package com.youmorry.expensetracker.user.presentation;
 
-import com.youmorry.expensetracker.application.UserService;
-import com.youmorry.expensetracker.domain.user.User;
-import com.youmorry.expensetracker.domain.user.UserId;
+import com.youmorry.expensetracker.user.application.UserService;
+import com.youmorry.expensetracker.user.domain.User;
+import com.youmorry.expensetracker.user.domain.UserId;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
