+++ backend/src/test/java/com/youmorry/expensetracker/domain/transaction/TransactionTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import com.youmorry.expensetracker.domain.category.CategoryId;
-import com.youmorry.expensetracker.domain.user.UserId;
+import com.youmorry.expensetracker.user.domain.UserId;
 import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDate;
