+++ backend/src/test/java/com/youmorry/expensetracker/infrastructure/repository/JdbcAnalyticsRepositoryTest.java
@@ -10,9 +10,9 @@ import com.youmorry.expensetracker.domain.transaction.Money;
 import com.youmorry.expensetracker.domain.transaction.NeedWantType;
 import com.youmorry.expensetracker.domain.transaction.Transaction;
 import com.youmorry.expensetracker.domain.transaction.TransactionRepository;
-import com.youmorry.expensetracker.domain.user.User;
-import com.youmorry.expensetracker.domain.user.UserId;
-import com.youmorry.expensetracker.domain.user.UserRepository;
+import com.youmorry.expensetracker.user.domain.User;
+import com.youmorry.expensetracker.user.domain.UserId;
+import com.youmorry.expensetracker.user.domain.UserRepository;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
