+++ backend/src/test/java/com/youmorry/expensetracker/infrastructure/repository/JdbcTransactionSearchRepositoryTest.java
@@ -9,9 +9,9 @@ import com.youmorry.expensetracker.domain.transaction.Transaction;
 import com.youmorry.expensetracker.domain.transaction.TransactionRepository;
 import com.youmorry.expensetracker.domain.transaction.TransactionSearchCriteria;
 import com.youmorry.expensetracker.domain.transaction.TransactionSearchRepository;
-import com.youmorry.expensetracker.domain.user.User;
-import com.youmorry.expensetracker.domain.user.UserId;
-import com.youmorry.expensetracker.domain.user.UserRepository;
+import com.youmorry.expensetracker.user.domain.User;
+import com.youmorry.expensetracker.user.domain.UserId;
+import com.youmorry.expensetracker.user.domain.UserRepository;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
